refactor: rename exceptions.py to match content
