+++ tests/security/test_jwt_validation_decoding.py
@@ -13,7 +13,7 @@ from fastapi.responses import JSONResponse
 from fastapi.testclient import TestClient
 
 import register_your_data_api.auth.authn as authn
-import register_your_data_api.exceptions
+import register_your_data_api.exception_handlers
 import tests.helpers.keys
 import tests.helpers.logs as logs
 import tests.helpers.mocking as mocking
@@ -33,7 +33,7 @@ async def lifespan(app: FastAPI) -> AsyncIterator[None]:
 
 
 app = FastAPI(title="Register Your Data", lifespan=lifespan)
-register_your_data_api.exceptions.add_exception_handlers(app)
+register_your_data_api.exception_handlers.add_exception_handlers(app)
 
 
 @app.get("/test_validate_and_decode_token")
